Add DropIcon and InputFiles for drag-and-drop functionality
DropIcon is an implementation of the visual feedback from the `upload
states` in the Input file task
(https://app.asana.com/0/520293273361225/708542321471201).
The actual drag and drop events and communication with external callback
functions is done with the InputFiles component.

InputFiles is handling the necessary browser events to enable drag and
drop into a DOM element and also calling callback functions passed to it
on adding files.